Move reload function to a seperate file
